Split user dto in two
Now we have one dto representing what we respond with and one representing what we need to create a new user.
