+++ src/controllers/interfaces/base-request.ts
@@ -1,6 +1,6 @@
-import { IUserDto } from '../../models';
+import { IUserResponseDto } from './';
 
 export interface IBaseRequest {
   headers: { [key: string]: string | string[], };
-  locals: { [user: string]: IUserDto, };
+  locals: { [user: string]: IUserResponseDto, };
 }
