+++ src/controllers/interfaces/index.ts
@@ -2,4 +2,5 @@ export { IBaseRequest } from './base-request';
 export { IBodyRequest } from './body-request';
 export { IBaseResponse } from './base-response';
 export { IBodyResponse } from './body-response';
-export { IUserDto } from './user-dto';
+export { IUserCreateDto } from './user-create-dto';
+export { IUserResponseDto } from './user-response-dto';
