+++ src/controllers/interfaces/user-create-dto.ts
@@ -1,0 +1,5 @@
+export interface IUserCreateDto {
+  id: number,
+  username: string,
+  token: string
+}
