--- src/controllers/interfaces/user-dto.ts
@@ -1,4 +1,0 @@
-export interface IUserDto {
-  id: number,
-  username: string,
-}
