+++ src/controllers/interfaces/user-response-dto.ts
@@ -1,0 +1,5 @@
+export interface IUserResponseDto {
+  id: number,
+  username: string,
+  token: string
+}
